+++ MinVer/MinVer.csproj
@@ -5,18 +5,25 @@
     <DevelopmentDependency>true</DevelopmentDependency>
     <IncludeBuildOutput>false</IncludeBuildOutput>
     <NoWarn>$(NoWarn);NU5100;NU5105</NoWarn>
+    <OutputType>Exe</OutputType>
     <PackageIconUrl>https://raw.githubusercontent.com/adamralph/min-ver/master/assets/min-ver.png</PackageIconUrl>
     <PackageLicenseUrl>https://github.com/adamralph/min-ver/blob/master/LICENSE</PackageLicenseUrl>
     <PackageProjectUrl>https://github.com/adamralph/min-ver</PackageProjectUrl>
     <PackageReleaseNotes>https://github.com/adamralph/min-ver/milestones?state=closed</PackageReleaseNotes>
     <PublishRepositoryUrl>true</PublishRepositoryUrl>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="LibGit2Sharp" Version="0.25.3" PrivateAssets="All" />
+    <PackageReference Include="LibGit2Sharp" Version="0.25.3" IncludeAssets="compile" />
+    <PackageReference Include="McMaster.Extensions.CommandLineUtils" Version="2.2.5" IncludeAssets="compile" />
     <PackageReference Include="Microsoft.CodeQuality.Analyzers" Version="2.6.2" PrivateAssets="All" />
     <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.0.0-beta-63127-02" PrivateAssets="All" />
+    <PackageReference Update="Microsoft.NETCore.App" PrivateAssets="All" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MinVer.Lib\MinVer.Lib.csproj" PrivateAssets="All" />
   </ItemGroup>
 
   <ItemGroup>
@@ -24,7 +31,7 @@
     <None Remove="buildMultiTargeting\**\*" />
     <Content Include="build\**\*" PackagePath="build" />
     <Content Include="buildMultiTargeting\**\*" PackagePath="buildMultiTargeting" />
-    <Content Include="$(MSBuildThisFileDirectory)..\MinVer.Tasks\bin\Release\netcoreapp2.1\publish\**\*" PackagePath="minver-tasks" Visible="false" Link="hidden" />
+    <Content Include="$(OutputPath)\publish\**\*" PackagePath="minver" />
   </ItemGroup>
 
 </Project>
