+++ MinVer/Verbosity.cs
