+++ MinVerTests/MinVerTests.csproj
@@ -13,7 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MinVer\MinVer.csproj" />
+    <ProjectReference Include="..\MinVer.Lib\MinVer.Lib.csproj" />
   </ItemGroup>
 
 </Project>
