+++ minver-cli/minver-cli.csproj
@@ -13,7 +13,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\MinVer.Tasks\Program.cs" Link="Program.cs" />
+    <Compile Include="..\MinVer\Program.cs" Link="Program.cs" />
   </ItemGroup>
 
   <ItemGroup>
@@ -24,7 +24,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MinVer\MinVer.csproj" />
+    <ProjectReference Include="..\MinVer.Lib\MinVer.Lib.csproj" />
   </ItemGroup>
 
 </Project>
